+++ Util/OrderTotals.php
@@ -7,6 +7,11 @@ use Magento\Sales\Api\Data\OrderInterface;
 
 class OrderTotals
 {
+    public function getValueTotal(OrderInterface $order): float
+    {
+        return (float)$order->getSubtotal() - (float)$order->getDiscountAmount();
+    }
+
     public function getShippingTotal(OrderInterface $order): float
     {
         return (float)$order->getShippingAmount() - (float)$order->getShippingDiscountAmount();